+++ exts/image.py
@@ -3,8 +3,7 @@ Image extension for qrm
 ---
 Copyright (C) 2019-2021 classabbyamp, 0x5c
 
-This file is part of qrm2 and is released under the terms of
-the GNU General Public License, version 2.
+SPDX-License-Identifier: LiLiQ-Rplus-1.1
 """