+++ resources/study.py
@@ -3,8 +3,7 @@ A listing of hamstudy command resources
 ---
 Copyright (C) 2019-2021 classabbyamp, 0x5c
 
-This file is part of discord-qrmbot and is released under the terms of
-the GNU General Public License, version 2.
+SPDX-License-Identifier: LiLiQ-Rplus-1.1
 """