+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
 ## [Unreleased]
+### Changed
+- Changed the licence to LiLiQ-Rplus-1.1
 
 
 ## [2.7.3] - 2021-04-12